app/src/main/java/com/marvelapp/model/CharacterData.java
package com.marvelapp.model;

import java.util.List;

/**
 * Data model for CharacterData
 */
public class CharacterData {
    private String offset;
    private String limit;
    private int total;
    private int count;
    private List<MarvelCharacter> results = null;

    public String getOffset() {
        return offset;
    }

    public void setOffset(String offset) {
        this.offset = offset;
    }

    public String getLimit() {
        return limit;
    }

    public void setLimit(String limit) {
        this.limit = limit;
    }

    public int getTotal() {
        return total;
    }

    public void setTotal(int total) {
        this.total = total;
    }

    public int getCount() {
        return count;
    }

    public void setCount(int count) {
        this.count = count;
    }

    public List<MarvelCharacter> getResults() {
        return results;
    }

    public void setResults(List<MarvelCharacter> results) {
        this.results = results;
    }
}
